Service revamp WIP.

+++ Source/RhythmCodex.Lib/Beatmania/Streamers/BeatmaniaPs2NewChartEventStreamReader.cs
@@ -8,7 +8,7 @@ namespace RhythmCodex.Beatmania.Streamers;
 [Service]
 public class BeatmaniaPs2NewChartEventStreamReader : IBeatmaniaPs2NewChartEventStreamReader
 {
-    public IList<BeatmaniaPs2Event> Read(Stream stream, int length)
+    public IList<BeatmaniaPs2Event> Read(Stream stream, long length)
     {
         throw new System.NotImplementedException();
     }
